+++ react_ui/src/API/axios.jsx
@@ -1,0 +1,5 @@
+import axios from 'axios';
+
+export default axios.create({
+    baseURL: 'http://localhost:8080/api/v1'
+});
