+++ react_ui/src/App.jsx
@@ -1,21 +1,21 @@
 import { BrowserRouter as Router, Routes, Route, Navigate } from 'react-router-dom'; // Import Navigate
 import PrivateRoute from './PrivateRoute'; // Import the PrivateRoute component
 
-import SignUpForm from './SignUpForm';
+import SignUpForm from './Login/SignUpForm';
 import HomePage from './HomePage';
-import Login from './Login';
+import Login from './Login/Login';
 
 function App() {
   return (
     <Router>
-        <Routes>
-          <Route path="/signup" element={<SignUpForm />} />
-          <Route path="/login" element={<Login />} />
-          <Route path="/" element={<Navigate to="/login" />} />
-          <Route element={<PrivateRoute />}>
-            <Route path="/home" element={<HomePage />} />
-          </Route>
-        </Routes>
+      <Routes>
+        <Route path="/signup" element={<SignUpForm />} />
+        <Route path="/login" element={<Login />} />
+        <Route path="/" element={<Navigate to="/login" />} />
+        <Route element={<PrivateRoute />}>
+          <Route path="/home" element={<HomePage />} />
+        </Route>
+      </Routes>
     </Router>
   );
 }
