+++ react_ui/src/Login/LogoutComponent.jsx
@@ -1,18 +1,15 @@
 import React from 'react';
-import axios from 'axios';
-import { FiLogOut} from 'react-icons/fi';
+import axios from "../API/axios.jsx";
+import { FiLogOut } from 'react-icons/fi';
 
 function LogoutComponent() {
   const handleLogout = async () => {
     try {
-      const axiosInstance = axios.create({
-        baseURL: 'http://localhost:8080',
+      await axios.post('/auth/register', {
         headers: {
           Authorization: `Bearer ${localStorage.getItem('jwtToken')}`
         }
       });
-      
-      await axiosInstance.post('/api/v1/auth/logout');
       localStorage.removeItem('jwtToken'); // Remove the token from local storage
       window.location.href = '/login'; // Redirect to homepage or login page
     } catch (error) {
